fix: only comment welcome message on non-bot user PRs (#24)
Currently, the welcome message is commented to every PR, even ones that
were opened by renovate. The message itself does not make sense on PRs
opened by bots and this PR fixes that behavior.

It also fixes a link at the end of the message to the location of the
workflow file that generated the message.