fix: validate the dictionaries correctly (#186)
# Description

We saw failing workflows complaining about

```
Checking for orphaned words in dictionary: .config/dictionaries/project.txt
The following word in the .config/dictionaries/project.txt dictionary is not being used: (boto)
The following word in the .config/dictionaries/project.txt dictionary is not being used: (botocore)
```

Not removing the `dictionaryDefinitions` resolves the problem, but I
don't know, why. It seems that CSpell is working differently. Having
`dictionaryDefinitions` without `dictionaries` shouldn't make a
difference as `dictionaries` activates the usage of the definitions.